feat: interface UserService
